full refactoring to app
