+++ src/components/CartItem.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import clsx from 'clsx';
 import { useDispatch } from 'react-redux';
-import { addItem, CartItem, minusItem, removeItem } from '../redux/slices/cartSlice';
+import { CartItem } from '../redux/cart/types';
+import { addItem, minusItem, removeItem } from '../redux/cart/slice';
 
 type CartItemProps = {
   id: string;
