+++ src/components/Header.tsx
@@ -2,8 +2,8 @@ import { Link, useLocation } from 'react-router-dom';
 import logoSvg from '../assets/img/pizza-logo.svg';
 import Search from './Search';
 import { useSelector } from 'react-redux';
-import { selectCart } from '../redux/slices/cartSlice';
 import React from 'react';
+import { selectCart } from '../redux/cart/selectors';
 
 function Header() {
   const { items, totalPrice } = useSelector(selectCart);
