+++ src/components/Sort.tsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { useDispatch } from 'react-redux';
-import { setSort, Sort, sortPropertyEnum } from '../redux/slices/filterSlice';
+import { setSort } from '../redux/filter/slice';
+import { Sort, sortPropertyEnum } from '../redux/filter/types';
 
 type SortItem = {
   name: string;
