+++ src/redux/cart/types.ts
@@ -1,0 +1,14 @@
+export type CartItem = {
+  id: string;
+  title: string;
+  price: number;
+  imageUrl: string;
+  type: string;
+  size: number;
+  count: number;
+};
+
+export interface CartSliceState {
+  totalPrice: number;
+  items: CartItem[];
+}
