+++ src/redux/pizza/types.ts
@@ -1,0 +1,20 @@
+export type Pizza = {
+  id: string;
+  title: string;
+  price: number;
+  imageUrl: string;
+  sizes: number[];
+  types: number[];
+  rating: number;
+};
+
+export enum Status {
+  LOADING = 'loading',
+  SUCCESS = 'success',
+  ERROR = 'error',
+}
+
+export interface PizzaSliceState {
+  items: Pizza[];
+  status: Status;
+}
