--- src/redux/slices/pizzaSlice.ts
@@ -1,71 +1,0 @@
-import { RootState } from './../store';
-import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
-import axios from 'axios';
-
-type Pizza = {
-  id: string;
-  title: string;
-  price: number;
-  imageUrl: string;
-  sizes: number[];
-  types: number[];
-  rating: number;
-};
-
-export enum Status {
-  LOADING = 'loading',
-  SUCCESS = 'success',
-  ERROR = 'error',
-}
-
-interface PizzaSliceState {
-  items: Pizza[];
-  status: Status;
-}
-
-const initialState: PizzaSliceState = {
-  items: [],
-  status: Status.LOADING, // loading | success | error
-};
-
-export const fetchPizzas = createAsyncThunk<Pizza[], Record<string, string>>(
-  'pizza/fetchPizzasStatus',
-  async (params) => {
-    const { queryString } = params;
-
-    const { data } = await axios.get<Pizza[]>(
-      `https://64074338862956433e6a09d1.mockapi.io/items?${queryString}`,
-    );
-    return data;
-  },
-);
-
-const pizzaSlice = createSlice({
-  name: 'pizza',
-  initialState,
-  reducers: {
-    setItems(state, action: PayloadAction<Pizza[]>) {
-      state.items = action.payload;
-    },
-  },
-  extraReducers: (builder) => {
-    builder.addCase(fetchPizzas.pending, (state) => {
-      state.status = Status.LOADING;
-      state.items = [];
-    });
-    builder.addCase(fetchPizzas.fulfilled, (state, action) => {
-      state.status = Status.SUCCESS;
-      state.items = action.payload;
-    });
-    builder.addCase(fetchPizzas.rejected, (state) => {
-      state.status = Status.ERROR;
-      state.items = [];
-    });
-  },
-});
-
-export const selectPizzaData = (state: RootState) => state.pizza;
-
-export const { setItems } = pizzaSlice.actions;
-
-export default pizzaSlice.reducer;
