+++ src/redux/store.ts
@@ -1,8 +1,8 @@
 import { useDispatch } from 'react-redux';
 import { configureStore } from '@reduxjs/toolkit';
-import filter from './slices/filterSlice';
-import cart from './slices/cartSlice';
-import pizza from './slices/pizzaSlice';
+import filter from './filter/slice';
+import pizza from './pizza/slice';
+import cart from './cart/slice';
 
 export const store = configureStore({
   reducer: {
