+++ src/utils/calcTotalPrice.ts
@@ -1,4 +1,4 @@
-import { CartItem } from '../redux/slices/cartSlice';
+import { CartItem } from '../redux/cart/types';
 
 export const calcTotalPrice = (items: CartItem[]) => {
   return items.reduce((sum, obj) => {
