+++ src/utils/getCartFromLS.ts
@@ -1,3 +1,4 @@
+import { CartItem } from '../redux/cart/types';
 import { calcTotalPrice } from './calcTotalPrice';
 
 export const getCartFromLS = () => {
@@ -6,7 +7,7 @@ export const getCartFromLS = () => {
   const totalPrice = calcTotalPrice(items);
 
   return {
-    items,
+    items: items as CartItem[],
     totalPrice,
   };
 };
